cloudscraper/requirements.txt
requests
cloudscraper
pycryptodome
Flask
pyinstaller
gevent
flask_compress